# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for 0x45",
        "title": "0x45 API",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://github.com/watzon/0x45/blob/main/LICENSE"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
        "/api/keys/request": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API Key"
                ],
                "summary": "Request a new API key",
                "operationId": "HandleRequestAPIKey",
                "parameters": [
                    {
                        "description": "Request a new API key",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.APIKeyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/services.APIKeyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/services.APIKeyResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "services.APIKeyRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "services.APIKeyResponse": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}